+++ src/exec/SimpleExecutor.ts
@@ -1,14 +1,9 @@
-import { parseYaml, stringifyYaml } from "obsidian";
+import { parseYaml } from "obsidian";
 
 // 创建一个简单的CoreExecutor实现，用于其他插件
 export class SimpleExecutor implements CoreExecutor {
 
     codes: Map<string, any> = new Map();
-    configPath: string;
-
-    constructor(configPath: string) {
-        this.configPath = configPath;
-    }
 
     excutable(type: string): boolean {
         return false;
@@ -24,17 +19,17 @@ export class SimpleExecutor implements CoreExecutor {
 
     async createConfigFile(name: string, content: any, started: boolean): Promise<boolean> {
         const configFile = `${pluto.self.settings.configPath}/${name}.yaml`;
-            if(started) {
-                if(!await app.vault.adapter.exists(configFile)) {
-                    const parsed = stringifyYaml(content);
-                    await app.vault.adapter.write(configFile, parsed);
-                    pluto.third.assets[name].yaml.set("config.yaml", content);
-                    return true;
-                }
+        if (started) {
+            if (!await app.vault.adapter.exists(configFile)) {
+                await app.vault.adapter.write(configFile, content);
+                let yaml = parseYaml(content);
+                pluto.third.assets[name].yaml.set(name, yaml);
+                return true;
             }
-            const ymlContent = await app.vault.adapter.read(configFile);
-            let yaml = parseYaml(ymlContent);
-            pluto.third.assets[name].yaml.set("config.yaml", yaml);
-            return false;
+        }
+        content = await app.vault.adapter.read(configFile);
+        let yaml = parseYaml(content);
+        pluto.third.assets[name].yaml.set(name, yaml);
+        return false;
     }
 }