+++ src/view/PlutoTextView.ts
@@ -47,6 +47,5 @@ export class PlutoTextView extends ItemView {
             this.state.file, //文件路径，可选
             this //组件上下文
         );
-        // pluto.third.dva.api.renderValue(text, contentEl);
     }
 }